+++ zcash_client_sqlite/src/lib.rs
@@ -1497,7 +1497,7 @@ impl<P: consensus::Parameters> WalletWrite for WalletDb<rusqlite::Connection, P>
     fn reserve_next_n_ephemeral_addresses(
         &mut self,
         account_id: Self::AccountId,
-        n: u32,
+        n: usize,
     ) -> Result<Vec<(TransparentAddress, TransparentAddressMetadata)>, Self::Error> {
         self.transactionally(|wdb| {
             wallet::transparent::ephemeral::reserve_next_n_ephemeral_addresses(wdb, account_id, n)
